# Performance improvments - fixes #10