By moving to maps based set implementation, performance is improved significantly.
current implementation:
 - loads in ~37s, 
 - spellchecking in ~1.7s (tested on included romeo & juliet example)

my implementation:
 - loads in ~1.1s (yes, that's a 37x faster)
 - spellchecking in ~0.8s (same test)

It also fixes a bug in generation of candidates and finds more of them.
Because you are matching on an empty set, which is an opaque type and can have many internal representations of an empty set, current implementation sometimes returns an empty list of candidates instead calling the catch-all case, which would return a one element list with the checked word. Whether returning the same word as a candidate makes any sens, is out of scope.

I also have a second implementation in a `binary_dict` branch, which can dump ETS table to a binary file, and when loading dictionary tries to populate ETS from it. This reduces loading to ~0.4s.