@ferigis 2017-12-31
Hi @X4lldux ,

Thanks for using/contributing to Sheldon! we really appreciate it!

The improvement is awesome, currently I don't have my laptop in order to test it myself but it seems very good!
I have one question, why the ets is `duplicated_bag`?

Once I can test it myself I will merge your PR, I will try today or tomorrow

Thanks!