@X4lldux 2017-12-31
@ferigis It's a low hanging fruit. I used `duplicate_bag` to avoid any unnecessary key checking because default type, `set`, can only have one instance of a key. Dictionaries already should have that property ;) Even if not, when creating a set from a list of words we ensure that there are no duplicates.