@ferigis 2017-12-31
Good catch! I like the PR. I will try to merge it today/tomorrow. Thanks!