@ferigis 2017-12-31
Hi! I tested your PR, everything is ok but `suggest_words` test is failing. It is because it is suggesting the misspelled word (as you mentioned).
It is easy to fix adding a filter in the list comprehension here
https://github.com/inaka/sheldon/pull/33/files#diff-9cd3c116be396c556d07d87be3a874e3L185

Could you replace that line by this ?
```erlang
[binary_to_list(Bin) || Bin <- Candidates, Bin =/= Word].
```