@X4lldux 2018-01-01
Assuming returning checked word as a valid suggestion is a bug, then actually the bug is here:
https://github.com/inaka/sheldon/pull/33/files#diff-9cd3c116be396c556d07d87be3a874e3L189
`know_sets/3` iterates through passed in sets, and when it goes through all of them without finding any suggestions, the that fallback line is executed returning the same word as a suggestion. Replacing `[Word]` with `[]` will make the tests pass again. In current implementation `EmptySet` rarely matched (due to matching on an opaque type) and fallback wasn't executed so the behavior didn't manifested.

Also, what's the intention with `Set1`? It's a single element set with the checked word - same as the fallback.