@ferigis 2018-01-01
you are right. I will create a new ticket for that.
Merged!
Thanks!!